requirements.txt

aiohttp==3.9.1
aiohttp-cors==0.7.0
python-dotenv==1.0.0
